ServiceInteractionIndicatorsTwo implementation
Completed the implementation of ServiceInteractionIndicatorsTwo for both
ASN.1 encoding/decoding and XML serialization. Also refactored the class
to be a subclass of SequenceBase.

Added implementation for ForwardServiceInteractionInd and
BackwardServiceInteractionInd in the process and placed factory methods
for them on the CAPParameterFactory API.

Added a generic enum XML formatter that can be registered with any enum type.
It serializes the enum by writing its name() in the "value" attribute, while
deserialization occurs completely during the newInstance() call.