Don't remove remoteGT even if RI=DPC+SSN
If the incoming SCCP message contains a calling address with RoutingIndicator set to DPC+SSN, but the address does not contain a PC, the stack inserts the MTP OPC (actual message origin) while also keeping the remote GT if present. The previous version removed the GT in such cases, which was problematic.